org-gcal-sync: update existing events if their ID can be found
This searches all files searched by `org-id-find` for the ID of each calendar
event, and only retrieves a new copy if the event is not present.

This also removes the behavior of erasing the `org-gcal-file-alist` files,
instead inserting events at the bottom of the file, so `org-gcal-header-alist`
is no longer necessary.