org-gcal-post-at-point: work even if ETag is absent
When the event ID was stored in the `ID` property, I decided to use
`ETag` as a backup to get greater assurance that the `ID` property that
might have already been stored in the event was known to GCal (and would
not 404). However, since commit 135cbe1205074704ffaf4cd1a372c1e77591cc0b
the event ID is stored in its own property, so the mere presence of an
`event-id` property means that the event is almost certainly known to
GCal.

This change makes the code more robust in case `ETag` is missing for
some reason, and makes generating headlines from `calendar-id` and
`event-id` which will then be populated by `org-gcal-post-at-point`
easier - it's no longer necessary to remember to add an `ETag` property,
which would be immediately overwritten anyway.